Organize resolution snapshot construction into focused files
